third_party/WebKit/LayoutTests/http/tests/inspector/elements/elements-linkify-attributes-expected.txt

Tests that you can click and hover on links in attributes.

    a link
<span class="highlight"><span class="webkit-html-tag">&lt;<span class="webkit-html-tag-name">img</span> <span class="webkit-html-attribute"><span class="webkit-html-attribute-name">id</span>=​"<span class="webkit-html-attribute-value">linkify-1</span>"</span> <span class="webkit-html-attribute"><span class="webkit-html-attribute-name">srcset</span>=​"<span class="webkit-html-attribute-value"><span class="devtools-link devtools-link-prevent-click">./​1x/​googlelogo_color_272x92dp.png</span> 1x, <span class="devtools-link devtools-link-prevent-click">./​2x/​googlelogo_color_272x92dp.png</span> 2x</span>"</span>&gt;</span>​</span>
<span class="highlight"><span class="webkit-html-tag">&lt;<span class="webkit-html-tag-name">img</span> <span class="webkit-html-attribute"><span class="webkit-html-attribute-name">id</span>=​"<span class="webkit-html-attribute-value">linkify-2</span>"</span> <span class="webkit-html-attribute"><span class="webkit-html-attribute-name">srcset</span>=​"<span class="webkit-html-attribute-value"><span class="devtools-link devtools-link-prevent-click">./​1x/​googlelogo_color_272x92dp.png</span></span>"</span>&gt;</span>​</span>
<span class="highlight"><span class="webkit-html-tag">&lt;<span class="webkit-html-tag-name">source</span> <span class="webkit-html-attribute"><span class="webkit-html-attribute-name">id</span>=​"<span class="webkit-html-attribute-value">linkify-3</span>"</span> <span class="webkit-html-attribute"><span class="webkit-html-attribute-name">media</span>=​"<span class="webkit-html-attribute-value">(min-width:​ 650px)​</span>"</span> <span class="webkit-html-attribute"><span class="webkit-html-attribute-name">srcset</span>=​"<span class="webkit-html-attribute-value"><span class="devtools-link devtools-link-prevent-click">./​kitten-large.png</span></span>"</span>&gt;</span>​</span>
<span class="highlight"><span class="webkit-html-tag">&lt;<span class="webkit-html-tag-name">source</span> <span class="webkit-html-attribute"><span class="webkit-html-attribute-name">id</span>=​"<span class="webkit-html-attribute-value">linkify-4</span>"</span> <span class="webkit-html-attribute"><span class="webkit-html-attribute-name">media</span>=​"<span class="webkit-html-attribute-value">(min-width:​ 465px)​</span>"</span> <span class="webkit-html-attribute"><span class="webkit-html-attribute-name">srcset</span>=​"<span class="webkit-html-attribute-value"><span class="devtools-link devtools-link-prevent-click">./​kitten-medium.png</span></span>"</span>&gt;</span>​</span>
<span class="highlight"><span class="webkit-html-tag">&lt;<span class="webkit-html-tag-name">source</span> <span class="webkit-html-attribute"><span class="webkit-html-attribute-name">id</span>=​"<span class="webkit-html-attribute-value">linkify-5</span>"</span> <span class="webkit-html-attribute"><span class="webkit-html-attribute-name">media</span>=​"<span class="webkit-html-attribute-value">(min-width:​ 400px)​</span>"</span> <span class="webkit-html-attribute"><span class="webkit-html-attribute-name">srcset</span>=​"<span class="webkit-html-attribute-value"><span class="devtools-link devtools-link-prevent-click">./​kitten-large.png</span> 2x</span>"</span>&gt;</span>​</span>
<span class="highlight"><span class="webkit-html-tag">&lt;<span class="webkit-html-tag-name">img</span> <span class="webkit-html-attribute"><span class="webkit-html-attribute-name">id</span>=​"<span class="webkit-html-attribute-value">linkify-6</span>"</span> <span class="webkit-html-attribute"><span class="webkit-html-attribute-name">src</span>=​"<span class="webkit-html-attribute-value"><span class="devtools-link devtools-link-prevent-click">./​kitten-small.png</span></span>"</span>&gt;</span>​</span>
<span class="highlight"><span class="webkit-html-tag">&lt;<span class="webkit-html-tag-name">a</span> <span class="webkit-html-attribute"><span class="webkit-html-attribute-name">id</span>=​"<span class="webkit-html-attribute-value">linkify-7</span>"</span> <span class="webkit-html-attribute"><span class="webkit-html-attribute-name">href</span>=​"<span class="webkit-html-attribute-value"><span class="devtools-link devtools-link-prevent-click" target="_blank">http:​/​/​www.google.com</span></span>"</span>&gt;</span>​<span class="webkit-html-text-node">a link</span>​<span class="webkit-html-tag">&lt;<span class="webkit-html-close-tag-name">/a</span>&gt;</span>​</span>
